requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.11.5
Django==1.11
django-admin-interface==0.5.0
django-celery==3.2.1
django-colorfield==0.1.14
django-crontab==0.7.1
django-flat-responsive==1.4.1
django-flat-theme==1.1.4
djangorestframework==3.6.2
enum34==1.1.6
futures==3.1.1
kombu==3.0.37
psycopg2==2.7.3.2
pypiwin32==219
pytz==2017.3
requests==2.6.0
six==1.11.0
suds==0.4
